Fixing unittest runs issue and review comments.
