+++ azurelinuxagent/common/osutil/default.py
@@ -16,7 +16,6 @@
 # Requires Python 2.6+ and Openssl 1.0+
 #
 
-import array
 import base64
 import datetime
 import errno
@@ -32,22 +31,21 @@ import socket
 import struct
 import sys
 import time
+from pwd import getpwall
 
-import azurelinuxagent.common.logger as logger
+import array
+
 import azurelinuxagent.common.conf as conf
+import azurelinuxagent.common.logger as logger
 import azurelinuxagent.common.utils.fileutil as fileutil
 import azurelinuxagent.common.utils.shellutil as shellutil
 import azurelinuxagent.common.utils.textutil as textutil
-
 from azurelinuxagent.common.exception import OSUtilError
 from azurelinuxagent.common.future import ustr
 from azurelinuxagent.common.utils.cryptutil import CryptUtil
 from azurelinuxagent.common.utils.flexible_version import FlexibleVersion
 from azurelinuxagent.common.utils.networkutil import RouteEntry, NetworkInterfaceCard
-
-from pwd import getpwall
-
-from azurelinuxagent.common.version import PY_VERSION_MAJOR, PY_VERSION_MINOR
+from azurelinuxagent.common.version import DISTRO_CODE_NAME
 
 __RULES_FILES__ = [ "/lib/udev/rules.d/75-persistent-net-generator.rules",
                     "/etc/udev/rules.d/70-persistent-net.rules" ]
@@ -301,16 +299,16 @@ class DefaultOSUtil(object):
     @staticmethod
     def is_cgroups_supported():
         """
-        Enabled by default; disabled in WSL/Travis (Disabled on Trusty only)
+        Enabled by default; disabled in WSL/Travis (Trusty only)
         """
         is_wsl = '-Microsoft-' in platform.platform()
-        supported = True  # By default supported in Travis
+        supported = True
+        base_fs_exists = os.path.exists(BASE_CGROUPS)
 
-        # Travis with python 2.6 fails (on Trusty based systems)
-        if PY_VERSION_MAJOR == 2 and PY_VERSION_MINOR == 6:
+        # Travis with python 2.6 fails on Trusty based systems
+        if DISTRO_CODE_NAME is "trusty":
             supported = False
 
-        base_fs_exists = os.path.exists(BASE_CGROUPS)
         return not is_wsl and base_fs_exists and supported
 
     @staticmethod
